requirements.txt
langchain
langchain-google-genai
langchain-community
google-search-results
python-dotenv